Basic multithreading and process/thread management

+++ src/hal/x86/include/memory.hpp
@@ -7,7 +7,7 @@
 
 #define align4k(addr)	((addr) & (~0xFFF))
 
-#define bytes_to_pages(bytes)	((bytes + memory::page_size_byte - 1) / memory::page_size_byte)
+#define bytes_to_pages(bytes)	(size_t)((bytes + memory::page_size_byte - 1) / memory::page_size_byte)
 
 extern const void *kernel_start;
 extern const void *kernel_end;
